% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitness.R
\docType{package}
\name{fitness-package}
\alias{fitness-package}
\title{Couple of functions to analyze JSON data from Google Fitness API}
\description{
The package is just to make the analysis more structured and repetitive
}

